iw3: Add --limit-resolution option
If the source resolution is lower than --resolution,
the depth resolution will be limited to the source resolution.

The ZoeDepth model already behaves as if --limit-resolution is enabled,
so this option is ignored. It is always equivalent to limit-resolution=True.

The DepthPro model only supports fixed-resolution input, so this option is ignored.